Started on refactoring the Service Registration/lookup
